+++ src/main/java/org/odk/g1/pankuru/Service/Service/HumainService/PersonnelService.java
@@ -2,9 +2,14 @@ package org.odk.g1.pankuru.Service.Service.HumainService;
 
 import java.util.List;
 import java.util.Optional;
+
+import org.odk.g1.pankuru.Entity.Humain.AdminCompagnie;
 import org.odk.g1.pankuru.Entity.Humain.Personnel;
+import org.odk.g1.pankuru.Repository.HumainRepo.AdminCompagnieRepo;
 import org.odk.g1.pankuru.Repository.HumainRepo.PersonnelRepo;
 import org.odk.g1.pankuru.Service.Interface.CrudService;
+import org.odk.g1.pankuru.Service.Service.CompagnieService;
+import org.odk.g1.pankuru.Service.Service.UserService;
 import org.springframework.stereotype.Service;
 
 import lombok.AllArgsConstructor;
@@ -15,8 +20,16 @@ import lombok.AllArgsConstructor;
 public class PersonnelService implements CrudService<Personnel, Long>{
 
     private final PersonnelRepo personnelRepo;
+    private final UserService userService;
+    private final AdminCompagnieRepo adminCompagnieRepo;
+    private final CompagnieService compagnieService;
+
     @Override
     public Personnel ajout(Personnel entity) {
+        Long adminCompagnieId = userService.getCurrentUsernameId();
+        Optional<AdminCompagnie> adminCompagnie = adminCompagnieRepo.findById(adminCompagnieId);
+        adminCompagnie.ifPresent(entity::setAdminCompagnie);
+
         return personnelRepo.save(entity);
     }
 
@@ -27,33 +40,56 @@ public class PersonnelService implements CrudService<Personnel, Long>{
 
     @Override
     public Optional<Personnel> trouverParId(Long id) {
-        return personnelRepo.findById(id);
+        Optional<Personnel> personnel = personnelRepo.findById(id);
+        if (personnel.isPresent()) {
+            if(compagnieService.getPersonnelByCompagnie().contains(personnel.get())){
+                return personnelRepo.findById(id);
+            }
+        }
+        return Optional.empty();
     }
 
     @Override
     public Personnel misAJour(Personnel entity, Long Id) {
-        Optional<Personnel> personnelExistant = personnelRepo.findById(Id);
-        if (personnelExistant.isPresent()) {
-            Personnel personnelAModifier = personnelExistant.get();
-            personnelAModifier.setNom(entity.getNom());
-            personnelAModifier.setPrenom(entity.getPrenom());
-            personnelAModifier.setEmail(entity.getEmail());
-            personnelAModifier.setPassword(entity.getPassword());
-            personnelAModifier.setNumeroDeTelephone(entity.getNumeroDeTelephone());
-            personnelAModifier.setDateEmbauche(entity.getDateEmbauche());
-            personnelAModifier.setPoste(entity.getPoste());
-            personnelAModifier.setEnService(entity.isEnService());
-            return personnelRepo.save(personnelAModifier);
-        } else {
-            throw new IllegalArgumentException("Le personnel avec l'ID " + entity.getId() + "n'existe pas.");
+        Optional<Personnel> personnel = personnelRepo.findById(Id);
+        if (personnel.isPresent()) {
+            if(compagnieService.getPersonnelByCompagnie().contains(personnel.get())){
+                Optional<Personnel> personnelExistant = personnelRepo.findById(Id);
+                if (personnelExistant.isPresent()) {
+                    Personnel personnelAModifier = modifierPersonnelExistant(entity, personnelExistant.get());
+
+                    Long adminCompagnieId = userService.getCurrentUsernameId();
+                    Optional<AdminCompagnie> adminCompagnie = adminCompagnieRepo.findById(adminCompagnieId);
+                    adminCompagnie.ifPresent(entity::setAdminCompagnie);
+
+                    return personnelRepo.save(personnelAModifier);
+                } else {
+                    throw new IllegalArgumentException("Le personnel avec l'ID " + entity.getId() + "n'existe pas.");
+                }
+            }
         }
+        return null;
     }
+
+    private static Personnel modifierPersonnelExistant(Personnel entity, Personnel personnelExistant) {
+        personnelExistant.setNom(entity.getNom());
+        personnelExistant.setPrenom(entity.getPrenom());
+        personnelExistant.setEmail(entity.getEmail());
+        personnelExistant.setPassword(entity.getPassword());
+        personnelExistant.setNumeroDeTelephone(entity.getNumeroDeTelephone());
+        personnelExistant.setDateEmbauche(entity.getDateEmbauche());
+        personnelExistant.setPoste(entity.getPoste());
+        personnelExistant.setEnService(entity.isEnService());
+        return personnelExistant;
+    }
+
 
     @Override
     public void supprimer(Long id) {
         Personnel personnelExistant = personnelRepo.findById(id).orElseThrow(() -> new IllegalArgumentException("Le personnel avec l'ID " + id + "n'existe pas."));
-        personnelRepo.delete(personnelExistant);
-    
+        if(compagnieService.getPersonnelByCompagnie().contains(personnelExistant)){
+            personnelRepo.delete(personnelExistant);
+        }
     }
 
 }
